+++ lib/Controller/FeedApiController.php
@@ -17,9 +17,11 @@ namespace OCA\News\Controller;
 
 use Exception;
 use OCA\News\Db\Feed;
+use OCA\News\Db\Filter;
 use OCA\News\Service\Exceptions\ServiceConflictException;
 use OCA\News\Service\Exceptions\ServiceNotFoundException;
 use OCA\News\Service\FeedServiceV2;
+use OCA\News\Service\FilterService;
 use OCA\News\Service\ItemServiceV2;
 use OCP\AppFramework\Http\JSONResponse;
 use \OCP\IRequest;
@@ -40,7 +42,8 @@ class FeedApiController extends ApiController
         ?IUserSession $userSession,
         private FeedServiceV2 $feedService,
         private ItemServiceV2 $itemService,
-        private LoggerInterface $logger
+        private LoggerInterface $logger,
+        private FilterService $filterService
     ) {
         parent::__construct($request, $userSession);
     }
@@ -223,6 +226,103 @@ class FeedApiController extends ApiController
             // ignore update failure
         } catch (Exception $ex) {
             $this->logger->debug('Could not update feed ' . $ex->getMessage());
+        }
+    }
+
+    /**
+     * Get the filter for a feed
+     *
+     * @param int $feedId
+     *
+     * @return array|JSONResponse
+     */
+    #[CORS]
+    #[NoCSRFRequired]
+    #[NoAdminRequired]
+    public function getFilter(int $feedId)
+    {
+        $filter = $this->filterService->findByFeedId($this->getUserId(), $feedId);
+
+        if ($filter === null) {
+            return [
+                'filter' => [
+                    'feedId'        => $feedId,
+                    'titleKeywords' => '',
+                    'bodyKeywords'  => '',
+                    'urlKeywords'   => '',
+                ]
+            ];
+        }
+
+        return ['filter' => $filter->toAPI()];
+    }
+
+    /**
+     * Save (create or update) a filter for a feed
+     *
+     * @param int         $feedId
+     * @param string|null $titleKeywords
+     * @param string|null $bodyKeywords
+     * @param string|null $urlKeywords
+     *
+     * @return array|JSONResponse
+     */
+    #[CORS]
+    #[NoCSRFRequired]
+    #[NoAdminRequired]
+    public function saveFilter(
+        int $feedId,
+        ?string $titleKeywords = null,
+        ?string $bodyKeywords = null,
+        ?string $urlKeywords = null
+    ) {
+        try {
+            $this->feedService->find($this->getUserId(), $feedId);
+            $filter = $this->filterService->findByFeedId($this->getUserId(), $feedId);
+
+            if ($filter === null) {
+                $filter = new Filter();
+                $filter->setFeedId($feedId);
+            }
+
+            $filter->setTitleKeywords($titleKeywords);
+            $filter->setBodyKeywords($bodyKeywords);
+            $filter->setUrlKeywords($urlKeywords);
+
+            if ($filter->getId() === null) {
+                $filter = $this->filterService->insert($filter);
+            } else {
+                $filter = $this->filterService->update($this->getUserId(), $filter);
+            }
+
+            return ['filter' => $filter->toAPI()];
+        } catch (ServiceNotFoundException $ex) {
+            return $this->error($ex, Http::STATUS_NOT_FOUND);
+        }
+    }
+
+    /**
+     * Delete the filter for a feed
+     *
+     * @param int $feedId
+     *
+     * @return array|JSONResponse
+     */
+    #[CORS]
+    #[NoCSRFRequired]
+    #[NoAdminRequired]
+    public function deleteFilter(int $feedId)
+    {
+        try {
+            $filter = $this->filterService->findByFeedId($this->getUserId(), $feedId);
+
+            if ($filter !== null) {
+                $this->filterService->delete($this->getUserId(), $filter->getId());
+            }
+
+            return [];
+        } catch (ServiceNotFoundException $ex) {
+            return $this->error($ex, Http::STATUS_NOT_FOUND);
         }
     }
 }