+++ src/types/FeedItem.ts
@@ -3,6 +3,7 @@ export type FeedItem = {
 	title: string
 	unread: boolean
 	starred: boolean
+	filtered: boolean
 	feedId: number
 	guidHash: string
 	pubDate: number